requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
iniconfig==2.0.0
lxml==5.2.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.1
requests==2.32.3
setuptools==69.1.1
urllib3==2.2.2
wheel==0.42.0
